Move setup-payments/* to setup-payment/*
To be consistent with setup-account-request singular.
